Install Eslint, Format Code, Add countdown when retry
