+++ src/sagas/ground.saga.js
@@ -1,9 +1,9 @@
-import { put, takeLatest, takeEvery, call, delay, take, retry } from 'redux-saga/effects'
-import { MAX_RETRIES, TIME_RETRY } from './../config'
-import groundService from '../services/ground.service'
-import { FETCHING_GROUNDS, FAILURE_GROUNDS, SUCCESS_GROUNDS, FETCHING_GROUND, SUCCESS_GROUND, FAILURE_GROUND, CLOSE_NOTIFICATION } from '../contants'
-import { NotiAction } from '../actions/notification.action'
+import { call, delay, put, take, takeLatest } from 'redux-saga/effects'
 import { GroundAction } from '../actions/ground.action'
+import { NotiAction } from '../actions/notification.action'
+import { CLOSE_NOTIFICATION, FETCHING_GROUND, FETCHING_GROUNDS } from '../contants'
+import groundService from '../services/ground.service'
+import { MAX_RETRIES, TIME_RETRY } from './../config'
 
 
 
@@ -22,24 +22,13 @@ function* getAll() {
     }
 }
 
-
-function* watchGetAll() {
-    yield takeLatest(FETCHING_GROUNDS, getAll)
-}
-
-
 function* get({ payload: { id } }) {
     try {
         const { data } = yield call(groundService.get, id)
-        yield put(GroundAction.getAll(data))
+        yield put(GroundAction.getSuccess(data))
     } catch (error) {
-        yield put({ type: FAILURE_GROUND })
+        yield put(GroundAction.getAllFail())
     }
-}
-
-
-function* watchGet() {
-    yield takeLatest(FETCHING_GROUND, get)
 }
 
 
@@ -48,14 +37,13 @@ function* retryApi(apiRequest, data) {
         console.log(i)
         try {
             const apiResponse = yield call(apiRequest, data);
-            console.log(apiResponse)
             return apiResponse;
         } catch (err) {
             if (err.response.status !== 404)
                 break;
             if (i < MAX_RETRIES) {
                 const timeRetry = i * 2 - 1;
-                yield put(NotiAction.open({ duration: timeRetry, message: `Retry after ${timeRetry}s` }))
+                yield put(NotiAction.open({ duration: timeRetry }))
                 yield delay(TIME_RETRY * timeRetry);
                 yield take(CLOSE_NOTIFICATION)
             }
@@ -64,9 +52,7 @@ function* retryApi(apiRequest, data) {
     throw new Error('API request failed');
 }
 
-
-export default {
-    // watchGetAll,
-    getAll,
-    watchGet
-}
+export const watchGrounds = [
+    takeLatest(FETCHING_GROUND, get),
+    getAll()
+]
