+++ src/sagas/index.js
@@ -1,10 +1,9 @@
 import { all } from 'redux-saga/effects'
-import groundSaga from "./ground.saga";
+import { watchGrounds } from "./ground.saga";
 
 
 export default function* rootSaga() {
     yield all([
-        groundSaga.getAll(),
-        groundSaga.watchGet()
+        ...watchGrounds
     ])
 }
